Move QueryStats to cli package
